Flytter IverksettResultat ut til domene. Splitter tilsand-pakken, sånn at tilstand er en egen pakke. Rettelser
