+++ src/main/kotlin/no/nav/familie/ef/iverksett/journalføring/DistribuerVedtaksbrevTask.kt
@@ -1,8 +1,8 @@
 package no.nav.familie.ef.iverksett.journalføring
 
 import com.fasterxml.jackson.module.kotlin.readValue
-import no.nav.familie.ef.iverksett.lagretilstand.DistribuerVedtaksbrevResultat
-import no.nav.familie.ef.iverksett.lagretilstand.LagreTilstandService
+import no.nav.familie.ef.iverksett.domene.DistribuerVedtaksbrevResultat
+import no.nav.familie.ef.iverksett.tilstand.lagre.LagreTilstandService
 import no.nav.familie.kontrakter.felles.objectMapper
 import no.nav.familie.prosessering.AsyncTaskStep
 import no.nav.familie.prosessering.TaskStepBeskrivelse
@@ -19,7 +19,8 @@ import java.util.*
                      triggerTidVedFeilISekunder = 15 * 60L,
                      beskrivelse = "Distribuerer vedtaksbrev.")
 class DistribuerVedtaksbrevTask(val journalpostClient: JournalpostClient,
-                                val lagreTilstandService: LagreTilstandService) : AsyncTaskStep {
+                                val lagreTilstandService: LagreTilstandService
+) : AsyncTaskStep {
 
     private val logger: Logger = LoggerFactory.getLogger(this::class.java)
 
@@ -27,7 +28,8 @@ class DistribuerVedtaksbrevTask(val journalpostClient: JournalpostClient,
         val taskData = objectMapper.readValue<DistribuerVedtaksbrevTaskData>(task.payload)
         val bestillingId = journalpostClient.distribuerBrev(taskData.journalpostId)
         lagreTilstandService.lagreDistribuerVedtaksbrevResultat(behandlingId = taskData.behandlingId,
-                                                                DistribuerVedtaksbrevResultat(bestillingId = bestillingId))
+                                                                DistribuerVedtaksbrevResultat(bestillingId = bestillingId)
+        )
         logger.info("Distribuer vedtaksbrev journalpost=[${taskData.journalpostId}] for behandling=[${taskData.behandlingId}] med bestillingId=[$bestillingId]")
     }
