+++ src/main/kotlin/no/nav/familie/ef/iverksett/journalføring/JournalførVedtaksbrevTask.kt
@@ -1,8 +1,8 @@
 package no.nav.familie.ef.iverksett.journalføring
 
+import no.nav.familie.ef.iverksett.domene.JournalpostResultat
 import no.nav.familie.ef.iverksett.hentIverksett.tjeneste.HentIverksettService
-import no.nav.familie.ef.iverksett.lagretilstand.JournalpostResultat
-import no.nav.familie.ef.iverksett.lagretilstand.LagreTilstandService
+import no.nav.familie.ef.iverksett.tilstand.lagre.LagreTilstandService
 import no.nav.familie.kontrakter.felles.dokarkiv.Dokumenttype
 import no.nav.familie.kontrakter.felles.dokarkiv.v2.ArkiverDokumentRequest
 import no.nav.familie.kontrakter.felles.dokarkiv.v2.Dokument
@@ -25,7 +25,8 @@ import java.util.*
 class JournalførVedtaksbrevTask(val hentIverksettService: HentIverksettService,
                                 val journalpostClient: JournalpostClient,
                                 val taskRepository: TaskRepository,
-                                val lagreTilstandService: LagreTilstandService) : AsyncTaskStep {
+                                val lagreTilstandService: LagreTilstandService
+) : AsyncTaskStep {
 
     override fun doTask(task: Task) {
         val behandlingId = UUID.fromString(task.payload)
@@ -46,7 +47,8 @@ class JournalførVedtaksbrevTask(val hentIverksettService: HentIverksettService,
 
         lagreTilstandService.lagreJournalPostResultat(behandlingId = behandlingId,
                                                       JournalpostResultat(journalpostId = journalpostId,
-                                                                          bestillingId = journalpostId))
+                                                                          bestillingId = journalpostId)
+        )
 
         lagDistribuerVedtaksbrevTask(behandlingId, journalpostId)
     }
